feat: Implement initial notification service
