+++ config/default.yaml
@@ -32,7 +32,13 @@ traefik:
 haproxy:
   use_tls: true
 notification_services:
-  gitlab:
+  chat.factorial.io:
+    type: mattermost
+    host: https://chat.factorial.io
+    channel: 17rf175n67f4mdae161fjgzzor
+    token: mattermost # Override with env SCOTTY__NOTIFICATION_SERVICES__CHAT_FACTORIAL_IO__TOKEN
+  source.factorial.io:
+    type: gitlab
     host: https://source.factorial.io
     user: deploybot
-    token: deploybot $ # Override with env SCOTTY__NOTIFICATION_SERVICES__GITLAB__TOKEN
+    token: deploybot # Override with env SCOTTY__NOTIFICATION_SERVICES__SOURCE_FACTORIAL_IO__TOKEN
