+++ src/api/handlers/blueprints.rs
@@ -3,7 +3,7 @@ use std::collections::HashMap;
 use axum::{debug_handler, extract::State, response::IntoResponse, Json};
 use serde::{Deserialize, Serialize};
 
-use crate::{app_state::SharedAppState, settings::AppBlueprint};
+use crate::{app_state::SharedAppState, settings::app_blueprint::AppBlueprint};
 
 #[derive(Debug, Clone, Serialize, Deserialize, utoipa::ToSchema, utoipa::ToResponse)]
 pub struct AppBlueprintList {
