+++ src/apps/app_data.rs
@@ -1,8 +1,7 @@
 #![allow(dead_code)]
 
-use std::{collections::HashMap, fmt::Display};
+use std::collections::HashMap;
 
-use bollard::secret::ContainerStateStatusEnum;
 use chrono::TimeDelta;
 use serde::{Deserialize, Serialize};
 use utoipa::{ToResponse, ToSchema};
@@ -51,19 +50,6 @@ impl Serialize for AppTtl {
             AppTtl::Forever => serializer.serialize_unit_variant("AppTtl", 2, "Forever"),
         }
     }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct GitlabContext {
-    pub project_id: u32,
-    pub mr_id: u32,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum NotificationReceiver {
-    Log,
-    Webhook,
-    Gitlab(GitlabContext),
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, ToSchema, ToResponse)]
@@ -152,51 +138,14 @@ impl AppSettings {
     }
 }
 
-use crate::settings::{AppBlueprintMap, Apps};
+pub use bollard::models::ContainerStateStatusEnum as ContainerStatus;
+
+use crate::{
+    notification_types::NotificationReceiver, settings::app_blueprint::AppBlueprintMap,
+    settings::config::Apps,
+};
 
 use super::create_app_request::CustomDomainMapping;
-
-#[derive(Debug, Serialize, Deserialize, Clone, ToSchema, ToResponse, PartialEq)]
-pub enum ContainerStatus {
-    Empty,
-    Created,
-    Restarting,
-    Running,
-    Paused,
-    Removing,
-    Exited,
-    Dead,
-}
-
-impl From<ContainerStateStatusEnum> for ContainerStatus {
-    fn from(status: ContainerStateStatusEnum) -> Self {
-        match status {
-            ContainerStateStatusEnum::EMPTY => ContainerStatus::Empty,
-            ContainerStateStatusEnum::CREATED => ContainerStatus::Created,
-            ContainerStateStatusEnum::RESTARTING => ContainerStatus::Restarting,
-            ContainerStateStatusEnum::RUNNING => ContainerStatus::Running,
-            ContainerStateStatusEnum::PAUSED => ContainerStatus::Paused,
-            ContainerStateStatusEnum::REMOVING => ContainerStatus::Removing,
-            ContainerStateStatusEnum::EXITED => ContainerStatus::Exited,
-            ContainerStateStatusEnum::DEAD => ContainerStatus::Dead,
-        }
-    }
-}
-
-impl Display for ContainerStatus {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ContainerStatus::Empty => write!(f, "empty"),
-            ContainerStatus::Created => write!(f, "created"),
-            ContainerStatus::Restarting => write!(f, "restarting"),
-            ContainerStatus::Running => write!(f, "running"),
-            ContainerStatus::Paused => write!(f, "paused"),
-            ContainerStatus::Removing => write!(f, "removing"),
-            ContainerStatus::Exited => write!(f, "exited"),
-            ContainerStatus::Dead => write!(f, "dead"),
-        }
-    }
-}
 
 #[derive(Debug, Serialize, Deserialize, Clone, ToSchema, ToResponse)]
 pub struct ContainerState {
@@ -213,7 +162,7 @@ pub struct ContainerState {
 impl Default for ContainerState {
     fn default() -> Self {
         ContainerState {
-            status: ContainerStatus::Empty,
+            status: ContainerStatus::EMPTY,
             id: None,
             service: "".to_string(),
             domain: None,
@@ -231,9 +180,9 @@ impl ContainerState {
     }
 
     pub fn is_running(&self) -> bool {
-        self.status == ContainerStatus::Running
-            || self.status == ContainerStatus::Created
-            || self.status == ContainerStatus::Restarting
+        self.status == ContainerStatus::RUNNING
+            || self.status == ContainerStatus::CREATED
+            || self.status == ContainerStatus::RESTARTING
     }
 
     pub fn running_since(&self) -> Option<TimeDelta> {
@@ -342,7 +291,7 @@ fn count_state(services: &[ContainerState], required: ContainerStatus) -> usize 
 }
 
 fn get_app_status_from_services(services: &[ContainerState]) -> AppStatus {
-    let count_running_services = count_state(services, ContainerStatus::Running);
+    let count_running_services = count_state(services, ContainerStatus::RUNNING);
     match count_running_services {
         0 => AppStatus::Stopped,
         x if x == services.len() => AppStatus::Running,
