+++ src/docker/find_apps.rs
@@ -111,7 +111,7 @@ pub async fn inspect_app(
     let settings = match get_app_settings(docker_compose_path).await {
         Ok(result) => Some(result),
         Err(err) => {
-            error!("{}", err);
+            info!("{}", err);
             None
         }
     };
