+++ src/main.rs
@@ -5,6 +5,7 @@ mod docker;
 mod http;
 mod init_telemetry;
 mod notification;
+mod notification_types;
 mod onepassword;
 mod settings;
 mod state_machine;
