+++ src/notification/mod.rs
@@ -1,0 +1,5 @@
+pub mod gitlab;
+pub mod log;
+pub mod mattermost;
+pub mod notify;
+pub mod webhook;
